Using TokenIcon on tooltip edge
